% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_lm.R
\name{output.lm}
\alias{output.lm}
\title{A ggplot2 function}
\usage{
output.lm(x)
}
\arguments{
\item{intercept}{pulls out intercept}

\item{slope}{pulls out slope}

\item{intercept_SE}{pulls out SE for intercept}

\item{slope_SE}{pulls out SE for slope}

\item{intercept_p}{pulls out p value for intercept}

\item{slope_p}{pulls out p value for slope}
}
\description{
This function allows you to quickly use a pre-customised theme.
}
\examples{
output.lm(model.name)
}
\keyword{lm,}
\keyword{summary}
