% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_shrub.R
\name{theme_shrub}
\alias{theme_shrub}
\title{A ggplot2 function}
\usage{
theme_shrub()
}
\arguments{
\item{theme}{Makes the theme}
}
\description{
This function allows you to quickly use a pre-customised theme, no angled x axis labels.
}
\examples{
ggplot() + geom_point(...) + theme_QHI()
}
\keyword{theme}
