% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_QHI.R
\name{theme_QHI}
\alias{theme_QHI}
\title{A ggplot2 function}
\usage{
theme_QHI()
}
\arguments{
\item{theme}{Makes the theme}
}
\description{
This function allows you to quickly use a pre-customised theme, clean theme with angled x axis labels.
}
\examples{
ggplot() + geom_point(...) + theme_QHI()
}
\keyword{theme}
